requirements.txt
# Core ML libraries
tensorflow==2.10.0
keras==2.10.0

# Web application
streamlit==1.22.0

# Data manipulation
pandas==1.5.3
numpy==1.23.5

# Image processing
Pillow==9.4.0
opencv-python==4.7.0.72

# Visualization
matplotlib==3.7.1
seaborn==0.12.2

# Model evaluation
scikit-learn==1.2.2

# Utilities
tqdm==4.65.0
joblib==1.2.0

# Image augmentation
albumentations==1.3.0

# Deep learning extras
efficientnet==1.1.1

# Notebook support (optional)
ipykernel==6.22.0
jupyter==1.0.0

